+++ apps/hash-frontend/src/shared/layout/plain-layout.tsx
@@ -1,25 +1,15 @@
-import { TextField } from "@hashintel/design-system";
-import { Autocomplete, Box, Modal, useTheme } from "@mui/material";
-import { bindPopover, usePopupState } from "material-ui-popup-state/hooks";
+import { useTheme } from "@mui/material";
 import Head from "next/head";
 import NextNProgress from "nextjs-progressbar";
 import { FunctionComponent, ReactNode } from "react";
-import { useKeys } from "rooks";
 
 import { isProduction } from "../../lib/config";
+import { CommandBar } from "../command-bar";
 
 export const PlainLayout: FunctionComponent<{
   children?: ReactNode;
 }> = ({ children }) => {
   const { palette } = useTheme();
-  const popupState = usePopupState({
-    popupId: "kbar",
-    variant: "popover",
-  });
-
-  useKeys(["Meta", "k"], () => {
-    popupState.toggle();
-  });
 
   return (
     <>
@@ -33,29 +23,7 @@ export const PlainLayout: FunctionComponent<{
         options={{ showSpinner: false }}
         showOnShallow
       />
-      <Modal {...bindPopover(popupState)}>
-        <Box
-          maxWidth={560}
-          width="100vw"
-          height="100vh"
-          display="flex"
-          alignItems="center"
-          margin="0 auto"
-        >
-          <Autocomplete
-            options={[]}
-            openOnFocus
-            sx={{ width: "100%" }}
-            renderInput={(props) => (
-              <TextField
-                onBlur={() => popupState.close()}
-                autoFocus
-                {...props}
-              />
-            )}
-          />
-        </Box>
-      </Modal>
+      <CommandBar />
       {children}
     </>
   );
